server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-multipart==0.0.6
boto3==1.34.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
moto[s3]>=5.0.0,<6.0.0
httpx==0.27.0